Remove all Sendmessages

+++ Assets/Script/AutoBullet.cs
@@ -36,7 +36,7 @@ public class AutoBullet : StageObject {
 	
 	
 
-	public override void ApplyHealthDamage(int value){
+	public override void ApplyHealthDamage(float value){
 	
 		base.ApplyHealthDamage(value);
 		if (current_health <= 0) {
