+++ Assets/Script/Character.cs
@@ -286,7 +286,7 @@ public class Character : StageObject {
 		}
 	}
 
-	public override void ApplyHealthDamage(int value){
+	public override void ApplyHealthDamage(float value){
 		base.ApplyHealthDamage (value);
 		if (current_health <= 0) {
 			current_status = STATUS.DAMAGE;
